added profile avatar upload logic

+++ src/helpers/Upload.ts
@@ -1,29 +1,46 @@
+/* eslint-disable consistent-return */
+import { Request, Response, NextFunction } from 'express';
 import { unlink } from 'fs/promises';
-import multer from 'multer';
 import sharp from 'sharp';
-import { Request } from 'express'
 
-import multerConfig from '../config/multer';
+import Users from '../models/Users';
 
 export default class Upload {
-    static async profileAvatar(req: Request) {
-        // todo
-        /* if (req.file) {
-            const filename = `${req.file.filename}.jpg`;
+    static async profileAvatar(
+        req: Request,
+        res: Response,
+        next: NextFunction
+    ) {
+        try {
+            if (
+                (req.file && req.file.mimetype === 'image/jpeg') ||
+                req.file.mimetype === 'image/jpg' ||
+                req.file.mimetype === 'image/png'
+            ) {
+                const filename = `${req.file.filename}.jpeg`;
+                await sharp(req.file.path)
+                    .resize(200, 200)
+                    .toFormat('jpeg')
+                    .toFile(`src/public/uploads/avatars/${filename}`);
 
-            await sharp(req.file.path)
-                .resize(200, 200)
-                .toFormat('jpeg')
-                .toFile(`./public/uploads/avatars/${filename}`);
+                await unlink(req.file.path);
 
-            await unlink(req.file.path);
+                if (global.SESSION_USER.avatar !== 'avatar.png') {
+                    const oldProfileAvatarPath = `src/public/uploads/avatars/${global.SESSION_USER.avatar}`;
+                    await unlink(oldProfileAvatarPath);
+                }
 
-            await Users.updateAvatarName(
-                `${req.file.filename}.jpg`,
-                req.session.userID
-            );
-        } else {
-            return res.status(400).json({ error: 'Invalid file type.' });
-        }*/
+                global.SESSION_USER = await Users.updateAvatarName(
+                    `${filename}`,
+                    global.SESSION_USER.id
+                );
+
+                req.session.userID = global.SESSION_USER.id;
+
+                return null;
+            }
+        } catch (error) {
+            return next(error);
+        }
     }
 }
